requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-dotenv
chromadb
sentence-transformers
Pillow
